+++ classmap.php
@@ -36,11 +36,11 @@ return [
     'zy\log\driver\FileLogger' => ZY_PATH . '/log/driver/FileLogger.php',
     // WEB
     'zy\web\Application' => ZY_PATH . '/web/Application.php',
-    'zy\web\ErrorAction' => ZY_PATH . '/web/ErrorAction.php',
     'zy\web\ErrorHandler' => ZY_PATH . '/web/ErrorHandler.php',
     'zy\web\Request' => ZY_PATH . '/web/Request.php',
     'zy\web\Response' => ZY_PATH . '/web/Response.php',
     'zy\web\UrlManager' => ZY_PATH . '/web/UrlManager.php',
+    'zy\web\HttpException' => ZY_PATH . '/web/HttpException.php',
     'zy\web\UrlNormalizerRedirectException' => ZY_PATH . '/web/UrlNormalizerRedirectException.php',
     'zy\web\NotFoundHttpException' => ZY_PATH . '/web/NotFoundHttpException.php',
 ];
