--- web/ErrorAction.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- *
- * @author Bi Zhiming <evan2884@gmail.com>
- * @created 2016/11/22  下午6:44
- * @since 1.0
- */
-
-namespace zy\web;
-
-use zy\base\Component;
-
-class ErrorAction extends Component
-{
-
-}
